Address review: move roulette_select to probability.rs, group ant tour+cost into a struct
- roulette_select is a generic weighted-index selection helper, not
  ACO-specific — moves alongside probability.rs's other reusable
  RNG/probability helpers (bernoulli, sample_without_replacement).
- Replaces the parallel ant_tours/ant_costs Vecs (matched by index via
  .zip()) with a single Vec<Ant> where Ant groups tour+cost together.
